Removed redundant approval tests and merged ContractTests into a single one
