--- src/jvm/src/test/java/io/eventdriven/strictland/RawApprovalsExampleTests.shoppingCartConfirmed_withNonDeterministicId_isCompatible.approved.txt
@@ -1,1 +1,0 @@
-{"shoppingCartId":"guid_1","clientId":"anonymised","confirmedAt":"2024-01-01T12:00:00Z"}
