--- src/jvm/src/test/java/io/eventdriven/strictland/RawApprovalsExampleTests.shoppingCartConfirmed_withOnlyRequiredData_isCompatible.approved.txt
@@ -1,1 +1,0 @@
-{"shoppingCartId":"00000000-0000-0000-0000-000000000001","clientId":null,"confirmedAt":"2024-01-01T12:00:00Z"}
